fix: logger in about
